pyproject.toml
[project]
name = "ai-uncomplicated"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "blobfile>=3.0.0",
    "datasets~=3.2.0",
    "datatrove>=0.4.0",
    "fsspec>=2024.6.1",
    "huggingface-hub~=0.26.2",
    "lm-eval>=0.4.8",
    "matplotlib==3.9.2",
    "msgspec>=0.19.0",
    "multiprocess==0.70.16",
    "omegaconf>=2.3.0",
    "orjson>=3.10.16",
    "pynvml>=12.0.0",
    "pyyaml~=6.0",
    "randomname>=0.2.1",
    "regex~=2024.11.6",
    "requests~=2.32.3",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.5.1",
    "scipy>=1.15.2",
    "seaborn==0.13.2",
    "sentencepiece~=0.1.99",
    "tiktoken>=0.9.0",
    "tokenizers>=0.21.1",
    "torch==2.5.0+cu121",
    "tqdm~=4.67.0",
    "transformers>=4.50.3",
    "uv>=0.6.16",
    "viztracer>=1.0.3",
    "wandb>=0.19.10",
    "xformers>=0.0.28.post2",
]

[dependency-groups]
ci = [
    "pre-commit>=4.0.1",
]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
